requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
bashlex
bracex
cachetools
certifi
cffi
charset-normalizer
cibuildwheel
cryptography
dataclasses-json
Django
django-allauth
django-jazzmin
filelock
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
h11
httpcore
httplib2
httpx
idna
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
marshmallow
multidict
mypy-extensions
numpy
oauthlib
ollama
orjson
packaging
pillow
platformdirs
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
pyparsing
python-decouple
python-dotenv
PyYAML
requests
requests-oauthlib
requests-toolbelt
rsa
sniffio
SQLAlchemy
sqlparse
tenacity
typing-inspect
typing_extensions
tzdata
uritemplate
urllib3
yarl
reportlab